TextVerse/backend/requirements.txt
# backend/requirements.txt

langchain
langchain-community
langchain-experimental
fastapi
uvicorn[standard]
python-multipart
pypdf
faiss-cpu
sentence-transformers
huggingface_hub
python-dotenv